Chore : 마이페이지 관련 컴포넌트 통합
MyPageHeader, MyPageUserInfo 컴포넌트를 한 개의 MyPage 컴포넌트 안에서 같이 관리하기로 하였습니다.
